add pkcs11 build option to the makefile, add a stub pkclient to avoid forcing CGO onto people
